% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{data}
\name{AILped}
\alias{AILped}
\title{Example AIL pedigree}
\format{A matrix with five columns: individual id, mom, dad, sex (0
for females and 1 for males) and generation.}
\source{
Derived from the pedF8 dataset in the QTLRel package,
\url{http://cran.r-project.org/web/packages/QTLRel/}
}
\usage{
data(AILped)
}
\description{
Example matrix describing the pedigree for advanced intercross lines
}
\examples{
data(AILped)
x <- sim_from_pedigree(AILped)
}
\keyword{datasets}

