% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_parent}
\alias{create_parent}
\title{Create a parent object}
\usage{
create_parent(L, allele = 1L)
}
\arguments{
\item{L}{chromosome length in cM}

\item{allele}{vector of integers for alleles, of length 1 or 2}
}
\value{
A list with two components, for the individual's two
chromosomes.  Each is a list with two components: an integer vector
of alleles in chromosome intervals, and a numeric vector of
locations of the right-endpoints of those intervals; these two
vectors should have the same length.
}
\description{
Create a parent object
}
\examples{
create_parent(100, 1)
create_parent(100, 1:2)
}
\seealso{
\code{\link{cross}}, \code{\link{meiosis}}
}
\keyword{datagen}

