% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geno2array}
\alias{geno2array}
\title{Convert multi-allele genotype matrix to an 3-dimensional array.}
\usage{
geno2array(geno, n_founders)
}
\arguments{
\item{geno}{Matrix of genotypes, as output by
\code{\link{convert2geno}}}

\item{n_founders}{Number of founder genotypes; if not provided,
inferred from contents.}
}
\value{
A 3-dimensional array, individual x marker x allele
}
\description{
Convert a multi-allele genotype matrix as output by
\code{\link{convert2geno}} to a 3-dimensional array, of dimension
\code{n_ind} x \code{n_markers} x 2, with the last dimension
corresponding to the two alleles; we can't recover phase
information, though.
}
\examples{
# simulate 8-way RIL pedigree
tab <- sim_ril_pedigree(ngen=20, parents=1:8)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab, L=100)
# marker map
map <- seq(0, 100, by=10)
names(map) <- paste0("m", map)
# convert dat to integer genotypes
geno <- convert2geno(dat, map)
# convert to 3-d array
arr <- geno2array(geno, 8)
}
\seealso{
\code{\link{get_geno}}, \code{\link{convert2geno}}
}
\keyword{utilities}

