% Generated by roxygen2 (4.0.2): do not edit by hand
\name{meiosis}
\alias{meiosis}
\title{Simulate meiosis}
\usage{
meiosis(parent, m = 10, p = 0)
}
\arguments{
\item{parent}{An individual object, as output by
\code{\link{create_parent}} or \code{\link{cross}}}

\item{m}{interference parameter for chi-square model}

\item{p}{Proportion of chiasmata coming from no-interference process.}
}
\value{
A list with two components: an integer vector of alleles in
chromosome intervals, and a numeric vector of locations of the
right-endpoints of those intervals; these two vectors should have
the same length.
}
\description{
Output a random meiotic product from an input individual.
}
\examples{
ind <- create_parent(100, 1:2)
prod <- meiosis(ind)
}
\seealso{
\code{\link{create_parent}}, \code{\link{cross}}
}
\keyword{datagen}

