% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseL_mgi-data.R
\docType{data}
\name{mouseL_mgi}
\alias{mouseL_mgi}
\title{Mouse chromosome lengths}
\format{
A numeric vector with lengths in cM for the 20 mouse chromosomes.
}
\source{
Taken from Table 1 of Cox et al. (2009) A new standard
genetic map for the laboratory mouse. Genetics 182:1335-1344.
\url{https://doi.org/10.1534/genetics.109.105486}{doi:10.1534/genetics.109.105486}
}
\usage{
data(mouseL_mgi)
}
\description{
Mouse chromosome lengths in cM from the Mouse Genome Informatics
(MGI) standard map.
}
\examples{
data(mouseL_mgi)
}
\seealso{
mouseL_cox
}
\keyword{datasets}
