% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{plot_ind}
\alias{plot_ind}
\title{Plot an individual}
\usage{
plot_ind(ind, center, length = 30, width = 3, gap = 3, col = CCcolors(),
  border = "black", lend = 1, ljoin = 1, allborders = FALSE, ...)
}
\arguments{
\item{ind}{An individual object, as output by \code{\link{create.parent}}
or \code{\link{cross}}}

\item{center}{(x,y) vector for the center of the individual}

\item{length}{Length of chromosomes}

\item{width}{Width of chromsomes}

\item{gap}{Gap between chromsomes}

\item{col}{Vector of colors}

\item{border}{Color for border}

\item{lend}{Passed to \code{\link[graphics]{rect}}}

\item{ljoin}{Passed to \code{\link[graphics]{rect}}}

\item{allborders}{If TRUE, put borders around all segments}

\item{...}{Additional arguments passed to rect()}
}
\value{
None.
}
\description{
Add an individual, as a pair of chromosomes, to a plot
}
\examples{
mom <- create_parent(100, 1:2)
dad <- create_parent(100, 3:4)
kid <- cross(mom, dad)
plot(0,0, type="n", xlim=c(0, 100), ylim=c(0,100),
     xaxt="n", yaxt="n", xlab="", ylab="")
plot_ind(mom, c(25, 66), length=50)
plot_ind(dad, c(75, 66), length=50)
plot_ind(kid, c(50, 25), length=50)
}
\keyword{hplot}

