% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_crossovers}
\alias{sim_crossovers}
\title{Simulate crossover locations using the Stahl model}
\usage{
sim_crossovers(L, m = 10L, p = 0)
}
\arguments{
\item{L}{length of chr in cM}

\item{m}{Interference paramater (\code{m=0} is no interference)}

\item{p}{Proportion of chiasmata from no-interference mechanism
(\code{p=0} gives pure chi-square model)}
}
\value{
Numeric vector of crossover locations, in cM
}
\description{
Simulate crossover locations on a single meiotic product using the
Stahl model.
}
\details{
Chiasma locations are a superposition of two
processes: a proportion p exhibiting no interference, and a
proportion (1-p) following the chi-square model with interference
parameter m.  Crossover locations are derived by thinning the
chiasma locations with probability 1/2.
}
\examples{
x <- sim_crossovers(200, 10, 0)
x <- sim_crossovers(200, 10, 0.04)
}
\references{
Copenhaver, G. P., Housworth, E. A. and Stahl, F. W. (2002) Crossover
interference in arabidopsis.  \emph{Genetics} \bold{160}, 1631--1639.

Foss, E., Lande, R., Stahl, F. W. and Steinberg, C. M. (1993) Chiasma
interference as a function of genetic distance. \emph{Genetics}
\bold{133}, 681--691.

Zhao, H., Speed, T. P. and McPeek, M. S. (1995) Statistical analysis
of crossover interference using the chi-square model.  \emph{Genetics}
\bold{139}, 1045--1056.
}
\keyword{datagen}

