% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sim_from_pedigree}
\alias{sim_from_pedigree}
\title{Simulate genotypes for pedigree}
\usage{
sim_from_pedigree(pedigree, L = 100, xchr = FALSE, m = 10, p = 0,
  obligate_chiasma = FALSE)
}
\arguments{
\item{pedigree}{Matrix describing a pedigree, with first four
columns being individual ID, mom ID, dad ID, and sex (female as
\code{0}, male as \code{1}).}

\item{L}{Length of chromosome in cM.}

\item{xchr}{If TRUE, simulate X chromosome.}

\item{m}{Crossover interference parameter, for chi-square model
(m=0 corresponds to no interference).}

\item{p}{proption of crossovers coming from no-interference process}

\item{obligate_chiasma}{If TRUE, require an obligate chiasma on the
4-strand bundle at meiosis.}
}
\value{
A list with each component being the data for one
individual, as produced by the \code{\link{cross}} function.  Those
results are a list with two components, corresponding to the
maternal and paternal chromosomes. The chromosomes are represented
as lists with two components: an integer vector of alleles in
chromosome intervals, and a numeric vector of locations of the
right-endpoints of those intervals; these two vectors should have
the same length.
}
\description{
Simulate genotypes along one chromosome for a pedigree
}
\examples{
# simulate AIL pedigree
tab <- sim_ail_pedigree(12, 30)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab)
}
\seealso{
\code{\link{check_pedigree}},
\code{\link{sim_ril_pedigree}}, \code{\link{sim_ail_pedigree}},
\code{\link{sim_ril_pedigree}}
}
\keyword{datagen}

