% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_meiosis}
\alias{sim_meiosis}
\title{Simulate meiosis}
\usage{
sim_meiosis(parent, m = 10, p = 0, obligate_chiasma = FALSE, Lstar)
}
\arguments{
\item{parent}{An individual object, as output by
\code{\link{create_parent}} or \code{\link{cross}}}

\item{m}{interference parameter for chi-square model}

\item{p}{Proportion of chiasmata coming from no-interference process.}

\item{obligate_chiasma}{If TRUE, require an obligate chiasma on the
4-strand bundle at meiosis.}

\item{Lstar}{Adjusted chromosome length, if
\code{obligate_chiasma=TRUE}. Calculated if not provided.}
}
\value{
A list with alleles in chromosome intervals (as integers)
and locations of the right endpoints of those intervals.
}
\description{
Output a random meiotic product from an input individual.
}
\examples{
ind <- create_parent(100, 1:2)
prod <- sim_meiosis(ind)
}
\seealso{
\code{\link{create_parent}}, \code{\link{cross}}
}
\keyword{datagen}

