% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_ril_pedigree}
\alias{sim_ril_pedigree}
\title{Check a pedigree for errors}
\usage{
sim_ril_pedigree(ngen = 20, selfing = FALSE, parents = 1:8,
  firstind = max(parents) + 1)

sim_ril_pedigree(ngen = 20, selfing = FALSE, parents = 1:8,
  firstind = max(parents) + 1)
}
\arguments{
\item{pedigree}{Numeric matrix with four columns: ID, mom ID, dad
ID, sex. Sex is coded as \code{0}=female, \code{1}=male. There can
be additional columns, but they'll be ignored.}

\item{ngen}{Number of generations of inbreeding}

\item{selfing}{If TRUE, use selfing}

\item{parents}{Vector of length 2, 4, or 8, indicating the parents' IDs}

\item{firstind}{Positive integer to assign to the first child}
}
\value{
TRUE (invisibly) if everything is okay; otherwise gives an error.

A matrix with five columns: individual ID, mother ID, father ID, sex, and generation.
Founders have \code{0} for mother and father ID. Sex is coded 0 for female and 1 for male.
}
\description{
Perform a series of checks on the tabular data for a pedigree, checking for problems

Generate a pedigree for multi-way recombinant inbred lines (a table of individual, mom, dad, sex)
}
\examples{
tab <- sim_ril_pedigree(7)
check_pedigree(tab)
tab <- sim_ril_pedigree(7)
}
\keyword{datagen}
\keyword{utilities}

