% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcross-package.R
\docType{package}
\name{simcross-package}
\alias{simcross}
\alias{simcross-package}
\title{simcross: Simulate Experimental Crosses}
\description{
Simulate and plot general experimental crosses. The focus is on simulating genotypes with an aim towards flexibility rather than speed. Meiosis is simulated following the Stahl model, in which chiasma locations are the superposition of two processes: a proportion p coming from a process exhibiting no interference, and the remainder coming from a process following the chi-square model.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/simcross/assets/vignettes/simcross.html}{R/simcross User Guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kbroman.org/simcross/}
  \item \url{https://github.com/kbroman/simcross}
  \item Report bugs at \url{https://github.com/kbroman/simcross/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

}
\keyword{internal}
