% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AILped-data.R
\docType{data}
\name{AILped}
\alias{AILped}
\title{Example AIL pedigree}
\format{
A data frame with five columns: individual id, mom, dad, sex (0
for females and 1 for males) and generation.
}
\source{
Derived from the pedF8 dataset in the QTLRel package,
\url{https://cran.r-project.org/package=QTLRel}
}
\usage{
data(AILped)
}
\description{
Example matrix describing the pedigree for advanced intercross lines
}
\examples{
data(AILped)
x <- sim_from_pedigree(AILped)
}
\keyword{datasets}
