% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross.R
\name{CCcolors}
\alias{CCcolors}
\title{Collaborative Cross colors}
\usage{
CCcolors(palette = c("new", "original", "official"))
}
\arguments{
\item{palette}{Which version of the colors to use? (New or original)}
}
\value{
vector of eight colors
}
\description{
Get the vector of colors for the Collaborative Cross
}
\examples{
CCcolors()
}
\keyword{color}
