% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross.R
\name{CCcolors}
\alias{CCcolors}
\title{Collaborative Cross colors}
\usage{
CCcolors()
}
\value{
vector of eight colors
}
\description{
Get the vector of colors for the Collaborative Cross
}
\examples{
CCcolors()
}
\keyword{color}

