% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert2geno}
\alias{convert2geno}
\title{Convert continuous allele information into marker genotypes}
\usage{
convert2geno(xodat, map, founder_geno)
}
\arguments{
\item{xodat}{The sort of detailed genotype/crossover data generated by
\code{\link{sim_from_pedigree}}}

\item{map}{vector of marker locations}

\item{founder_geno}{Optional matrix (size \code{n_founders} x
\code{length(map)}) of founder genotypes coded as 1/2. If provided,
results are 1/2/3 genotypes}
}
\value{
If \code{founder_geno} is provided or there are just two
founders, the result is a numeric matrix of genotypes, individuals
x markers, with genotypes 1/2/3 codes for 11/12/22 genotypes.

If \code{founder_geno} is not provided and there are more than two
founders, the result is a 3-dimensional array, individuals x
markers x alleles, with the third dimensional corresponding to the
maternal and paternal allele.
}
\description{
Convert the continuous crossover location information produced by
sim_from_pedigree to marker genotypes
}
\examples{
# simulate AIL pedigree
tab <- sim_ail_pedigree(12, 30)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab)
# marker map (could also use sim.map in R/qtl)
map <- seq(0, 100, by=5)
names(map) <- paste0("marker", seq(along=map))
# convert data to marker genotypes
geno <- convert2geno(dat, map)

# simulate DO pedigree
tab <- sim_do_pedigree(8)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab)
# simulate founder snp alleles
fg <- matrix(sample(1:2, 8*length(map), repl=TRUE), nrow=8)
# for DO, need female & male founders (to deal with X chr)
fg <- rbind(fg, fg)
# convert dat to SNP genotypes
geno <- convert2geno(dat, map, fg)
# if fg not provided, result is a 3d array
genoarray <- convert2geno(dat, map)
}
\seealso{
\code{\link{get_geno}}, \code{\link{sim_from_pedigree}}
}
\keyword{utilities}

