% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert2geno}
\alias{convert2geno}
\title{Convert continuous allele information into marker genotypes}
\usage{
convert2geno(xodat, map)
}
\arguments{
\item{xodat}{The sort of detailed genotype/crossover data generated by
\code{\link{sim_from_pedigree}}}

\item{map}{vector of marker locations}
}
\value{
A numeric matrix of genotypes, individuals x markers. If
there are just two alleles, the genotypes are coded as 1/2/3. If
there are more than two alleles, a binary code is used.
}
\description{
Convert the continuous crossover location information produced by
sim_from_pedigree to marker genotypes
}
\examples{
# simulate AIL pedigree
tab <- sim_ail_pedigree(12, 30)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab)
# marker map (could also use sim.map in R/qtl)
map <- seq(0, 100, by=5)
names(map) <- paste0("marker", seq(along=map))
# convert data to marker genotypes
geno <- convert2geno(dat, map)
}
\seealso{
\code{\link{get_geno}}, \code{\link{sim_from_pedigree}}
}
\keyword{utilities}

