% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_meiosis.R
\name{create_parent}
\alias{create_parent}
\title{Create a parent object}
\usage{
create_parent(L, allele = 1)
}
\arguments{
\item{L}{chromosome length in cM}

\item{allele}{vector of integers for alleles, of length 1 or 2}
}
\value{
A list with two components, for the individual's two
chromosomes.  Each is a list with alleles in chromosome intervals
(as integers) and locations of the right endpoints of those
intervals.
}
\description{
Create a parent object
}
\examples{
create_parent(100, 1)
create_parent(100, 1:2)
}
\seealso{
\code{\link[=cross]{cross()}}, \code{\link[=sim_meiosis]{sim_meiosis()}}
}
\keyword{datagen}
