% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{create_parent}
\alias{create_parent}
\title{Create a parent object}
\usage{
create_parent(L, allele = 1)
}
\arguments{
\item{L}{chromosome length in cM}

\item{allele}{vector of alleles, of length 1 or 2}
}
\value{
a list with two components, for the two chromosomes.
Each is a matrix with two rows: locations of crossovers, and
the alleles in the segments
}
\description{
Create a parent object
}
\examples{
create_parent(100, 1)
create_parent(100, 1:2)
}
\keyword{datagen}

