% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cross}
\alias{cross}
\title{Cross two individuals}
\usage{
cross(mom, dad, m = 10, obligate.chiasma = FALSE, xchr = FALSE,
  male = FALSE)
}
\arguments{
\item{mom}{An individual object, as produced by
\code{\link{create_parent}} or \code{\link{cross}}}

\item{dad}{An individual object, as produced by
\code{\link{create_parent}} or \code{\link{cross}}}

\item{m}{interference parameter for chi-square model}

\item{obligate.chiasma}{If TRUE, simulate meiosis with an
obligate chiasma on the four-strand bundle}

\item{xchr}{If TRUE, simulate X chromosome}

\item{male}{If TRUE, simulate a male (matters only if
\code{xchr=TRUE})}
}
\value{
A matrix with two rows: locations of crossovers, and the
allele in each segment
}
\description{
Simulate the cross of two individuals to create a
single progeny
}
\examples{
mom <- create_parent(100, 1:2)
dad <- create_parent(100, 1:2)
child <- cross(mom, dad)
}
\seealso{
\code{\link{create_parent}}, \code{\link{meiosis}}
}
\keyword{datagen}

