% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cross}
\alias{cross}
\title{Cross two individuals}
\usage{
cross(mom, dad, m = 10, p = 0, xchr = FALSE, male = FALSE)
}
\arguments{
\item{mom}{An individual object, as produced by
\code{\link{create_parent}} or this function.}

\item{dad}{An individual object, as produced by
\code{\link{create_parent}} or this function.}

\item{m}{interference parameter for chi-square model}

\item{p}{proption of crossovers coming from no-interference process}

\item{xchr}{If TRUE, simulate X chromosome}

\item{male}{If TRUE, simulate a male (matters only if
\code{xchr=TRUE})}
}
\value{
A list with two components, for the individual's two
chromosomes.  Each is a data frame with two columns: alleles in
chromosome intervals (as integers), and locations of the
right endpoints of those intervals.
}
\description{
Simulate the cross of two individuals to create a
single progeny
}
\examples{
mom <- create_parent(100, 1:2)
dad <- create_parent(100, 1:2)
child <- cross(mom, dad)
}
\seealso{
\code{\link{create_parent}}, \code{\link{sim_meiosis}}
}
\keyword{datagen}

