% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{meiosis}
\alias{meiosis}
\title{Simulate meiosis}
\usage{
meiosis(parent, m = 10, obligate.chiasma = TRUE)
}
\arguments{
\item{parent}{An individual object, as output by
\code{\link{create_parent}} or \code{\link{cross}}}

\item{m}{interference parameter for chi-square model}

\item{obligate.chiasma}{If TRUE, simulate meiosis with an
obligate chiasma on the four-strand bundle}
}
\value{
A matrix with two rows: locations of crossovers, and the
allele in each segment
}
\description{
Output a random meiotic product from an input individual.
}
\examples{
ind <- create_parent(100, 1:2)
prod <- meiosis(ind)
}
\keyword{datagen}

