% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_do_pedigree}
\alias{sim_do_pedigree}
\title{Simulate diversity outcross pedigree}
\usage{
sim_do_pedigree(ngen = 12, npairs = 30, ccgen = rep(0, npairs),
  nkids_per = 5, design = c("nosib", "random"))
}
\arguments{
\item{ngen}{Number of generations of outbreeding}

\item{npairs}{Number of breeding pairs at each generation}

\item{ccgen}{Vector of same length as npairs, with the number of
generations for each CC line}

\item{nkids_per}{Number of offspring per pair for the last
generation}

\item{design}{How to choose crosses: either random but avoiding
siblings, or completely at random}
}
\value{
A matrix with six columns: individual ID, mother ID, father
ID, sex, generation, and 1/0 indicator for whether DO or pre-DO.
Founders have \code{0} for mother and father ID. Sex is coded 0 for
female and 1 for male.
}
\description{
Simulate a diversity outcross pedigree (a table of individual, mom,
dad, sex)
}
\details{
A diversity outcross is generated from a set of 8 inbred
lines. We need two individuals from each line (one female and one
male) as the order of the initial crosses will be randomized; for
example, sometimes the individual from line 1 will be a mother and
sometimes a father. The founders are numbered 1-8 for the females
from the 8 lines, and 9-16 for the corresponding males.

The diversity outcross is generated by first creating a panel of
partially-inbred 8-way RIL (the so-called pre-CC, for
pre-Collaborative Cross). The \code{ccgen} argument specifies the
number of inbreeding generations for each of the CC lines. We
generate a pre-CC line for each of the \code{npairs} breeding
pairs, and generate a sibling pair from each as the starting
material.

The subsequent \code{ngen} outbreeding generations then proceed by
crossing a male and female from the preceding generation (mated
completely at random, with \code{design="random"}, or avoiding
siblings, with \code{design="nosib"}). Each breeding pair gives a
single female and a single male to the next generation, except at
the last generation \code{nkids_per} offspring are mated, in equal
numbers male and female. (If \code{nkids_per} is an odd number, the
number of males and females in each sibship will differ by one,
alternating between sibships, with one additional female and then
one additional male.
}
\examples{
tab <- sim_do_pedigree(8, 30, rep(6, 30))
}
\seealso{
\code{\link{sim_from_pedigree}},
\code{\link{sim_ril_pedigree}}, \code{\link{sim_ail_pedigree}}
}
\keyword{datagen}

