% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_from_pedigree}
\alias{sim_from_pedigree}
\title{Simulate a pedigree}
\usage{
sim_from_pedigree(pedigree, L = 100, xchr = FALSE, m = 10,
  obligate.chiasma = TRUE)
}
\arguments{
\item{pedigree}{Matrix describing a pedigree, with first four
columns being individual ID, mom ID, dad ID, and sex (female as
\code{0}, male as \code{1}).}

\item{L}{Length of chromosome in cM.}

\item{xchr}{If TRUE, simulate X chromosome.}

\item{m}{Crossover interference parameter, for chi-square model
(m=0 corresponds to no interference).}

\item{obligate.chiasma}{If TRUE, 4-strand bundle has at least one
chiasma.}
}
\value{
A list whose each component is the data for one individual,
as produced by the \code{\link{cross}} function.  Those results are
a list of two matrices, corresponding to the maternal and paternal
chromosomes. The chromosomes are represented as a matrix with two
rows: a vector of positions, and the alleles at each position (and
in interval to left).
}
\description{
Simulate genotypes along one chromosome for a pedigree
}
\examples{
# simulate AIL pedigree
tab <- sim_ail_pedigree(12, 30)
# simulate data from that pedigree
dat <- sim_from_pedigree(tab)
}
\keyword{datagen}

