% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_ril_pedigree}
\alias{sim_ril_pedigree}
\title{Generate a ril pedigree}
\usage{
sim_ril_pedigree(ngen = 20, selfing = FALSE, parents = 1:8,
  firstind = max(parents) + 1)
}
\arguments{
\item{ngen}{Number of generations of inbreeding}

\item{selfing}{If TRUE, use selfing}

\item{parents}{Vector of length 2, 4, or 8, indicating the parents'
IDs}

\item{firstind}{Positive integer to assign to the first child}
}
\value{
A matrix with five columns: individual ID, mother ID,
father ID, sex, and generation.  Founders have \code{0} for mother
and father ID. Sex is coded 0 for female and 1 for male.
}
\description{
Generate a pedigree for multi-way recombinant inbred lines (a table
of individual, mom, dad, sex)
}
\examples{
tab <- sim_ril_pedigree(7)
}
\keyword{datagen}

